+++ src/experiments.ts
@@ -77,6 +77,15 @@ export const ALL_EXPERIMENTS = experiments({
     shortDescription: "Use Cloud Run API to list v2 functions",
     public: false,
   },
+  bypassfunctionsdeprecationcheck: {
+    shortDescription: "Bypass Functions check for old runtimes",
+    fullDescription:
+      "Bypasses the local check for whether a functions runtime is " +
+      "decommissioned. This does not, by itself, allow you to deploy a function with a " +
+      "decommissioned runtime, as there are server-side checks as well.",
+    public: false,
+    default: false,
+  },
 
   // Emulator experiments
   emulatoruisnapshot: {
@@ -182,6 +191,11 @@ export const ALL_EXPERIMENTS = experiments({
   fdcwebhooks: {
     shortDescription: "Enable Firebase Data Connect webhooks feature.",
     default: true,
+    public: false,
+  },
+  fdcrealtime: {
+    shortDescription: "Enable Firebase Data Connect realtime feature.",
+    default: false,
     public: false,
   },
 });