+++ CHANGELOG.md
@@ -1,1 +1,2 @@
-- Updated default Postgres version for Data Connect to 18.
+- Add validation to check if backendId exists in `firebase.json` when `--only` flag is used. (#10161)
+- Updated default Postgres version for Data Connect to 18. (#10234)